Move task fragment and view models under /tasks/<type>/
