+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/date/DateTaskFragment.kt
@@ -13,38 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.datacollection
+package com.google.android.ground.ui.datacollection.tasks.date
 
+import android.app.DatePickerDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.activityViewModels
 import com.google.android.ground.BR
-import com.google.android.ground.databinding.NumberTaskFragBinding
+import com.google.android.ground.databinding.DateTaskFragBinding
 import com.google.android.ground.ui.common.AbstractFragment
-import com.google.android.ground.ui.datacollection.TaskFragment.Companion.POSITION
-import com.google.android.ground.ui.editsubmission.NumberTaskViewModel
+import com.google.android.ground.ui.datacollection.DataCollectionViewModel
+import com.google.android.ground.ui.datacollection.TaskFragment
 import dagger.hilt.android.AndroidEntryPoint
+import java.util.*
 import kotlin.properties.Delegates
 
-/** Fragment allowing the user to answer questions to complete a task. */
 @AndroidEntryPoint
-class NumberTaskFragment : AbstractFragment(), TaskFragment<NumberTaskViewModel> {
+class DateTaskFragment : AbstractFragment(), TaskFragment<DateTaskViewModel> {
   private val dataCollectionViewModel: DataCollectionViewModel by activityViewModels()
-  override lateinit var viewModel: NumberTaskViewModel
+  override lateinit var viewModel: DateTaskViewModel
   override var position by Delegates.notNull<Int>()
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     if (savedInstanceState != null) {
-      position = savedInstanceState.getInt(POSITION)
+      position = savedInstanceState.getInt(TaskFragment.POSITION)
     }
   }
 
   override fun onSaveInstanceState(outState: Bundle) {
     super.onSaveInstanceState(outState)
-    outState.putInt(POSITION, position)
+    outState.putInt(TaskFragment.POSITION, position)
   }
 
   override fun onCreateView(
@@ -53,12 +54,35 @@ class NumberTaskFragment : AbstractFragment(), TaskFragment<NumberTaskViewModel>
     savedInstanceState: Bundle?
   ): View {
     super.onCreateView(inflater, container, savedInstanceState)
-    val binding = NumberTaskFragBinding.inflate(inflater, container, false)
+    val binding = DateTaskFragBinding.inflate(inflater, container, false)
 
-    viewModel = dataCollectionViewModel.getTaskViewModel(position) as NumberTaskViewModel
+    viewModel = dataCollectionViewModel.getTaskViewModel(position) as DateTaskViewModel
     binding.lifecycleOwner = this
     binding.setVariable(BR.viewModel, viewModel)
+    binding.setVariable(BR.fragment, this)
 
     return binding.root
+  }
+
+  fun showDateDialog() {
+    val calendar = Calendar.getInstance()
+    val year = calendar[Calendar.YEAR]
+    val month = calendar[Calendar.MONTH]
+    val day = calendar[Calendar.DAY_OF_MONTH]
+    val datePickerDialog =
+      DatePickerDialog(
+        requireContext(),
+        { _, updatedYear, updatedMonth, updatedDayOfMonth ->
+          val c = Calendar.getInstance()
+          c[Calendar.DAY_OF_MONTH] = updatedDayOfMonth
+          c[Calendar.MONTH] = updatedMonth
+          c[Calendar.YEAR] = updatedYear
+          viewModel.updateResponse(c.time)
+        },
+        year,
+        month,
+        day
+      )
+    datePickerDialog.show()
   }
 }
