+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/multiplechoice/OptionListItemDetailsLookup.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.datacollection
+package com.google.android.ground.ui.datacollection.tasks.multiplechoice
 
 import android.view.MotionEvent
 import android.view.View
