+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/number/NumberTaskFragment.kt
@@ -13,39 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.datacollection
+package com.google.android.ground.ui.datacollection.tasks.number
 
-import android.app.TimePickerDialog
 import android.os.Bundle
-import android.text.format.DateFormat
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.activityViewModels
 import com.google.android.ground.BR
-import com.google.android.ground.databinding.TimeTaskFragBinding
+import com.google.android.ground.databinding.NumberTaskFragBinding
 import com.google.android.ground.ui.common.AbstractFragment
-import com.google.android.ground.ui.editsubmission.TimeTaskViewModel
+import com.google.android.ground.ui.datacollection.DataCollectionViewModel
+import com.google.android.ground.ui.datacollection.TaskFragment
+import com.google.android.ground.ui.datacollection.TaskFragment.Companion.POSITION
 import dagger.hilt.android.AndroidEntryPoint
-import java.util.*
 import kotlin.properties.Delegates
 
+/** Fragment allowing the user to answer questions to complete a task. */
 @AndroidEntryPoint
-class TimeTaskFragment : AbstractFragment(), TaskFragment<TimeTaskViewModel> {
+class NumberTaskFragment : AbstractFragment(), TaskFragment<NumberTaskViewModel> {
   private val dataCollectionViewModel: DataCollectionViewModel by activityViewModels()
-  override lateinit var viewModel: TimeTaskViewModel
+  override lateinit var viewModel: NumberTaskViewModel
   override var position by Delegates.notNull<Int>()
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     if (savedInstanceState != null) {
-      position = savedInstanceState.getInt(TaskFragment.POSITION)
+      position = savedInstanceState.getInt(POSITION)
     }
   }
 
   override fun onSaveInstanceState(outState: Bundle) {
     super.onSaveInstanceState(outState)
-    outState.putInt(TaskFragment.POSITION, position)
+    outState.putInt(POSITION, position)
   }
 
   override fun onCreateView(
@@ -54,33 +54,12 @@ class TimeTaskFragment : AbstractFragment(), TaskFragment<TimeTaskViewModel> {
     savedInstanceState: Bundle?
   ): View {
     super.onCreateView(inflater, container, savedInstanceState)
-    val binding = TimeTaskFragBinding.inflate(inflater, container, false)
+    val binding = NumberTaskFragBinding.inflate(inflater, container, false)
 
-    viewModel = dataCollectionViewModel.getTaskViewModel(position) as TimeTaskViewModel
+    viewModel = dataCollectionViewModel.getTaskViewModel(position) as NumberTaskViewModel
     binding.lifecycleOwner = this
     binding.setVariable(BR.viewModel, viewModel)
-    binding.setVariable(BR.fragment, this)
 
     return binding.root
-  }
-
-  fun showTimeDialog() {
-    val calendar = Calendar.getInstance()
-    val hour = calendar[Calendar.HOUR]
-    val minute = calendar[Calendar.MINUTE]
-    val timePickerDialog =
-      TimePickerDialog(
-        requireContext(),
-        { _, updatedHourOfDay, updatedMinute ->
-          val c = Calendar.getInstance()
-          c[Calendar.HOUR_OF_DAY] = updatedHourOfDay
-          c[Calendar.MINUTE] = updatedMinute
-          viewModel.updateResponse(c.time)
-        },
-        hour,
-        minute,
-        DateFormat.is24HourFormat(requireContext())
-      )
-    timePickerDialog.show()
   }
 }
