+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/text/TextTaskViewModel.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.editsubmission
+package com.google.android.ground.ui.datacollection.tasks.text
 
 import android.content.res.Resources
-import com.google.android.ground.model.submission.TimeTaskData.Companion.fromDate
-import java.util.*
+import com.google.android.ground.model.submission.TextTaskData.Companion.fromString
+import com.google.android.ground.ui.editsubmission.AbstractTaskViewModel
 import javax.inject.Inject
 
-class TimeTaskViewModel @Inject constructor(resources: Resources) :
+class TextTaskViewModel @Inject constructor(resources: Resources) :
   AbstractTaskViewModel(resources) {
 
-  fun updateResponse(date: Date) {
-    setResponse(fromDate(date))
+  fun updateResponse(text: String) {
+    setResponse(fromString(text))
   }
 }
