+++ ground/src/main/java/com/google/android/ground/ui/editsubmission/TaskViewFactory.kt
@@ -24,8 +24,13 @@ import com.google.android.ground.BR
 import com.google.android.ground.R
 import com.google.android.ground.model.task.Task
 import com.google.android.ground.ui.common.ViewModelFactory
-import com.google.android.ground.ui.datacollection.DropAPinTaskViewModel
-import com.google.android.ground.ui.datacollection.PolygonDrawingViewModel
+import com.google.android.ground.ui.datacollection.tasks.date.DateTaskViewModel
+import com.google.android.ground.ui.datacollection.tasks.number.NumberTaskViewModel
+import com.google.android.ground.ui.datacollection.tasks.photo.PhotoTaskViewModel
+import com.google.android.ground.ui.datacollection.tasks.point.DropAPinTaskViewModel
+import com.google.android.ground.ui.datacollection.tasks.polygon.PolygonDrawingViewModel
+import com.google.android.ground.ui.datacollection.tasks.text.TextTaskViewModel
+import com.google.android.ground.ui.datacollection.tasks.time.TimeTaskViewModel
 import com.google.android.ground.ui.util.ViewUtil.assignGeneratedId
 import javax.inject.Inject
