+++ ground/src/main/res/layout/number_input_task.xml
@@ -23,7 +23,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.editsubmission.NumberTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.number.NumberTaskViewModel" />
   </data>
 
   <LinearLayout
