+++ ground/src/main/res/layout/number_task_frag.xml
@@ -22,7 +22,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.editsubmission.NumberTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.number.NumberTaskViewModel" />
   </data>
 
   <ScrollView
