+++ ground/src/main/res/layout/photo_input_task.xml
@@ -27,7 +27,7 @@
       type="com.google.android.ground.ui.editsubmission.EditSubmissionViewModel" />
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.editsubmission.PhotoTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.photo.PhotoTaskViewModel" />
   </data>
 
   <LinearLayout
