+++ ground/src/main/res/layout/question_task_frag.xml
@@ -22,7 +22,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.editsubmission.TextTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.text.TextTaskViewModel" />
   </data>
 
   <ScrollView
