+++ ground/src/main/res/layout/text_input_task.xml
@@ -23,7 +23,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.editsubmission.TextTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.text.TextTaskViewModel" />
   </data>
 
   <LinearLayout
