+++ ground/src/main/res/layout/time_input_task.xml
@@ -23,7 +23,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.editsubmission.TimeTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.time.TimeTaskViewModel" />
   </data>
 
   <LinearLayout
