+++ ground/src/test/java/com/google/android/ground/ui/editsubmission/TimeTaskViewModelTest.kt
@@ -17,6 +17,7 @@ package com.google.android.ground.ui.editsubmission
 
 import com.google.android.ground.BaseHiltTest
 import com.google.android.ground.model.submission.TimeTaskData.Companion.fromDate
+import com.google.android.ground.ui.datacollection.tasks.time.TimeTaskViewModel
 import com.google.common.truth.Truth.assertThat
 import com.sharedtest.TestObservers.observeUntilFirstChange
 import dagger.hilt.android.testing.HiltAndroidTest
